+++ src/fundamentalTypes/literalType.js
@@ -1,8 +1,38 @@
 import {$e} from '../utils/templateUtils.js';
 
+import bigintType from './bigintType.js';
 import booleanType from './booleanType.js';
+import nullType from './nullType.js';
 import numberType from './numberType.js';
 import stringType from './stringType.js';
+import undefinedType from './undefinedType.js';
+
+/**
+ * @typedef {bigint|boolean|null|number|string|undefined} LiteralValue
+ */
+
+/**
+ * @param {unknown} value
+ * @returns {import('../types.js').TypeObject}
+ */
+function getLiteralTypeObject (value) {
+  if (value === null) {
+    return nullType;
+  }
+  if (value === undefined) {
+    return undefinedType;
+  }
+  switch (typeof value) {
+  case 'bigint':
+    return bigintType;
+  case 'boolean':
+    return booleanType;
+  case 'number':
+    return numberType;
+  default:
+    return stringType;
+  }
+}
 
 /**
  * @type {import('../types.js').TypeObject}
@@ -13,17 +43,32 @@ const literalType = {
   valueMatch (x) {
     // const schema =
     //   /** @type {import('zodexy').SzLiteral<any>} */ (cfg?.schema);
-    return (typeof x === 'boolean' || typeof x === 'number' ||
+    return (x === null || x === undefined || typeof x === 'bigint' ||
+      typeof x === 'boolean' || typeof x === 'number' ||
       typeof x === 'string'); // && (!schema || schema.value === x);
   },
   toValue (s) {
-    const value = s === 'true'
-      ? true
-      : s === 'false'
-        ? false
-        : s.charAt(0) === '"'
-          ? s.slice(1, -1)
+    let value;
+    switch (s) {
+    case 'true':
+      value = true;
+      break;
+    case 'false':
+      value = false;
+      break;
+    case 'null':
+      value = null;
+      break;
+    case 'undefined':
+      value = undefined;
+      break;
+    default:
+      value = s.charAt(0) === '"'
+        ? s.slice(1, -1)
+        : s.endsWith('n')
+          ? BigInt(s.slice(0, -1))
           : Number(s);
+    }
     return {value};
   },
   // Todo: Fix all the following methods up to `setValue` to work with children
@@ -31,11 +76,7 @@ const literalType = {
     return /** @type {HTMLTextAreaElement} */ ($e(root, 'input,textarea'));
   },
   setValue ({root, value}) {
-    const typeObject = typeof value === 'boolean'
-      ? booleanType
-      : typeof value === 'number'
-        ? numberType
-        : stringType;
+    const typeObject = getLiteralTypeObject(value);
     typeObject.setValue?.({root, value});
   },
   getValue ({root, stateObj}) {
@@ -59,17 +100,12 @@ const literalType = {
     const {specificSchemaObject} = arg;
     const {
       values
-    } = /** @type {import('zodexy').SzLiteral<(boolean|number|string)[]>} */ (
+    } = /** @type {import('zodexy').SzLiteral<LiteralValue[]>} */ (
       specificSchemaObject
     );
-    const literalValue = arg.value !== undefined ? arg.value : values[0];
+    const literalValue = Object.hasOwn(arg, 'value') ? arg.value : values[0];
 
-    // Todo: BigInt, null, undefined
-    const typeObject = typeof literalValue === 'boolean'
-      ? booleanType
-      : typeof literalValue === 'number'
-        ? numberType
-        : stringType;
+    const typeObject = getLiteralTypeObject(literalValue);
     const specificLiteralEditUI = typeObject.editUI({
       ...arg,
       value: literalValue